+++ config/routes/base.rb
@@ -43,9 +43,7 @@ resources :projects do
   member do
     get 'select'
     get 'settings_for'
-    get 'stats'
     get 'recently_created_stats'
-    get 'per_relationship_recent_stats/:relationship', action: :per_relationship_recent_stats, as: :per_relationship_recent_stats
   end
 end